cisTEMParameters & cistem_star_file_reader & defines.h 	*Change UNSIGNED_INTEGER to INTEGER_UNSIGNED see issue #287
Image::CalculateCtfImage

	*Add option to apply the CTF coherence envelope. Default false.'

Image::Compute1DPowerSpectrumCurve
	*Add option to average amplitudes not intensities for Guinier analysis
		bool average_amplitudes_not_intensities

make_template_result/ match_template

	*Change default edge exclusion radius to 1/4 box size, rather than 1/2. Box should be padded to 2x particle radius meaning a full particle (less delocalized info) is still in the image with only 1/4 excluded.
	*Note that in match_template it was hardcoded to 50 pixels

match_template

	*Change the range of the stored histogram to better match realistic outputs
	*Enable symmetry from CLI
	*Enable number of threads to be specified, which is important for saturating the GPU, especially for smaller images. (CLI and GUI)
	*variable temp_threshold now initialized to 1.0. This only matters when compiled w/o MKL
	*Fix threshold calc
	*Edge pixels to ignore was hardcoded to 50, change to box_size/4 to match new default in make_template_results

gui/MatchTemplatePanel

	*Changes to take threads from runprofile and pass to match_template

ax_cuda.m4

	*Add compatibility with Ampere arch

src/gpu

	*Changes to core functions to work with half-precision values to accelerate template matching further.

Add Image::Compute1DAmplitudeSpectrumCurve for naming consistency which calls Compute1DAmplitudeSpectrumCurve with average_amplitudes_not_intensities=true

When threads are requested in match_template, send a warning (SendInfo) and override to max_threads=1 instead of crashing.

+++ src/gpu/Histogram.h
@@ -23,14 +23,19 @@ public:
 	dim3 threadsPerBlock_accum_array;
 	dim3 gridDims_accum_array;
 
+//	float* histogram;		bool is_allocated_histogram; // histogram_n_bins in size;
 	float* histogram;		bool is_allocated_histogram; // histogram_n_bins in size;
+
 	size_t size_of_temp_hist;
 	float* cummulative_histogram;
 
 	int histogram_n_bins; //
-	float histogram_min;
-	float histogram_max;
-	float histogram_step;
+//	float histogram_min;
+//	float histogram_max;
+//	float histogram_step;
+	__half histogram_min;
+	__half histogram_max;
+	__half histogram_step;
 
 	int max_padding;
